// client/src/pages/systeminfo/systeminfo.interface.ts
/**
 * systeminfo.state 参数类型
 *
 * @export
 * @interface ISysteminfoState
 */
export interface ISysteminfoState {
  systeminfo: any;
  infoProps: Array<{ key: string; title: string }>;
}
/**
 * systeminfo.props 参数类型
 *
 * @export
 * @interface ISysteminfoProps
 */
export interface ISysteminfoProps {}